android: clean up
